Contribution table

One row per outcome and disease with raw and adjusted impacts, the disease's burden (Shapley share of the expected loss, including any interaction share) and, if economics is given, the attributed gap and its monetary value.

Usage

contribution_table(result, economics = NULL)

Arguments

result

A deconflate() result.

economics

Optional list with observed and unit_value (see value_losses()).

Value

A data frame.

Examples

contribution_table(deconflate(example_uk_dairy_2022(), method = "published"),
                   uk_dairy_2022_economics())
#>      outcome disease        raw     adjusted        change         main
#> 1      yield      CO 0.00000000 0.0000000000            NA 0.000000e+00
#> 2      yield      DA 0.04040000 0.0236694153 -4.141234e-01 7.100825e-04
#> 3      yield     DYS 0.04050000 0.0293974969 -2.741359e-01 5.879499e-04
#> 4      yield     FAS 0.07330000 0.0733000000  0.000000e+00 7.330000e-03
#> 5      yield     GIN 0.03280000 0.0328000000  0.000000e+00 6.888000e-03
#> 6      yield     LAM 0.05540000 0.0483062939 -1.280452e-01 1.449189e-02
#> 7      yield     MAS 0.04570000 0.0376265428 -1.766621e-01 1.128796e-02
#> 8      yield     MET 0.03950000 0.0240913756 -3.900918e-01 2.409138e-03
#> 9      yield      MF 0.00410000 0.0009988662 -7.563741e-01 7.990929e-05
#> 10     yield     NEO 0.04200000 0.0420000000  0.000000e+00 6.300000e-03
#> 11     yield     PTB 0.05900000 0.0443065737 -2.490411e-01 3.101460e-03
#> 12     yield      RP 0.07380000 0.0610014912 -1.734215e-01 3.050075e-03
#> 13     yield     SCK 0.03050000 0.0191867269 -3.709270e-01 4.221080e-03
#> 14 fertility      CO 0.11260000 0.1112817871 -1.170704e-02 1.001536e-02
#> 15 fertility      DA 0.00000000 0.0000000000            NA 0.000000e+00
#> 16 fertility     DYS 0.06960000 0.0601977177 -1.350903e-01 1.203954e-03
#> 17 fertility     FAS 0.00000000 0.0000000000            NA 0.000000e+00
#> 18 fertility     GIN 0.01200000 0.0120000000 -3.330669e-16 2.520000e-03
#> 19 fertility     LAM 0.12470000 0.1188955682 -4.654717e-02 3.566867e-02
#> 20 fertility     MAS 0.00000000 0.0000000000            NA 0.000000e+00
#> 21 fertility     MET 0.04740000 0.0409885292 -1.352631e-01 4.098853e-03
#> 22 fertility      MF 0.00000000 0.0000000000            NA 0.000000e+00
#> 23 fertility     NEO 0.07210000 0.0721000000  0.000000e+00 1.081500e-02
#> 24 fertility     PTB 0.05790000 0.0385501384 -3.341945e-01 2.698510e-03
#> 25 fertility      RP 0.02740000 0.0167647793 -3.881467e-01 8.382390e-04
#> 26 fertility     SCK 0.01500000 0.0054953314 -6.336446e-01 1.208973e-03
#> 27   culling      CO 0.00000000 0.0000000000            NA 0.000000e+00
#> 28   culling      DA 0.31384146 0.2189054562 -3.024967e-01 6.567164e-03
#> 29   culling     DYS 0.14205130 0.1189588327 -1.625643e-01 2.379177e-03
#> 30   culling     FAS 0.00000000 0.0000000000            NA 0.000000e+00
#> 31   culling     GIN 0.00000000 0.0000000000            NA 0.000000e+00
#> 32   culling     LAM 0.25564849 0.2255177810 -1.178599e-01 6.765533e-02
#> 33   culling     MAS 0.21306880 0.1692001905 -2.058894e-01 5.076006e-02
#> 34   culling     MET 0.17385764 0.1226083366 -2.947774e-01 1.226083e-02
#> 35   culling      MF 0.20567043 0.1555226953 -2.438257e-01 1.244182e-02
#> 36   culling     NEO 0.09889327 0.0988932697  0.000000e+00 1.483399e-02
#> 37   culling     PTB 0.19637306 0.1343726145 -3.157279e-01 9.406083e-03
#> 38   culling      RP 0.00000000 0.0000000000            NA 0.000000e+00
#> 39   culling     SCK 0.15726427 0.0966549023 -3.853982e-01 2.126408e-02
#>    interaction       burden       share          gap      value
#> 1            0 0.000000e+00 0.000000000   0.00000000  0.0000000
#> 2            0 7.100825e-04 0.011745142   6.60320396  1.9954882
#> 3            0 5.879499e-04 0.009725005   5.46746832  1.6522689
#> 4            0 7.330000e-03 0.121242105  68.16318906 20.5989157
#> 5            0 6.888000e-03 0.113931189  64.05293946 19.3567983
#> 6            0 1.449189e-02 0.239703550 134.76307130 40.7254001
#> 7            0 1.128796e-02 0.186708919 104.96910567 31.7216637
#> 8            0 2.409138e-03 0.039848419  22.40306941  6.7702076
#> 9            0 7.990929e-05 0.001321742   0.74309309  0.2245627
#> 10           0 6.300000e-03 0.104205356  58.58500561 17.7043887
#> 11           0 3.101460e-03 0.051299803  28.84112075  8.7157867
#> 12           0 3.050075e-03 0.050449858  28.36327543  8.5713818
#> 13           0 4.221080e-03 0.069818911  39.25269702 11.8621650
#> 14           0 1.001536e-02 0.145008175   3.75669401 14.7585481
#> 15           0 0.000000e+00 0.000000000   0.00000000  0.0000000
#> 16           0 1.203954e-03 0.017431546   0.45159512  1.7741366
#> 17           0 0.000000e+00 0.000000000   0.00000000  0.0000000
#> 18           0 2.520000e-03 0.036486014   0.94523493  3.7134499
#> 19           0 3.566867e-02 0.516431598  13.37907667 52.5610406
#> 20           0 0.000000e+00 0.000000000   0.00000000  0.0000000
#> 21           0 4.098853e-03 0.059345558   1.53745196  6.0400338
#> 22           0 0.000000e+00 0.000000000   0.00000000  0.0000000
#> 23           0 1.081500e-02 0.156585812   4.05663324 15.9368893
#> 24           0 2.698510e-03 0.039070581   1.01219270  3.9765003
#> 25           0 8.382390e-04 0.012136508   0.31441776  1.2352216
#> 26           0 1.208973e-03 0.017504208   0.45347755  1.7815319
#> 27           0 0.000000e+00 0.000000000   0.00000000  0.0000000
#> 28           0 6.567164e-03 0.033239927   0.14806119  1.9771499
#> 29           0 2.379177e-03 0.012042285   0.05364016  0.7162893
#> 30           0 0.000000e+00 0.000000000   0.00000000  0.0000000
#> 31           0 0.000000e+00 0.000000000   0.00000000  0.0000000
#> 32           0 6.765533e-02 0.342439827   1.52533569 20.3687227
#> 33           0 5.076006e-02 0.256923794   1.14442014 15.2821287
#> 34           0 1.226083e-02 0.062058636   0.27642886  3.6913205
#> 35           0 1.244182e-02 0.062974682   0.28050923  3.7458080
#> 36           0 1.483399e-02 0.075082758   0.33444244  4.4660106
#> 37           0 9.406083e-03 0.047609216   0.21206656  2.8318520
#> 38           0 0.000000e+00 0.000000000   0.00000000  0.0000000
#> 39           0 2.126408e-02 0.107628873   0.47941316  6.4018916